# analysis/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DocumentIngestResult": {
        "additionalProperties": false,
        "properties": {
          "doc_type": {
            "title": "Doc Type",
            "type": "string"
          },
          "fields": {
            "items": {
              "type": "string"
            },
            "title": "Fields",
            "type": "array"
          },
          "household_id": {
            "title": "Household Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "const": "accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "household_id",
          "name",
          "doc_type",
          "fields",
          "status"
        ],
        "title": "DocumentIngestResult",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "enum": [
              "needs_review",
              "not_found"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "errors"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FindingEvidence": {
        "additionalProperties": false,
        "properties": {
          "check": {
            "enum": [
              "verified",
              "mismatch",
              "unconfirmed",
              "conflicted",
              "not_checked"
            ],
            "title": "Check",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "headline": {
            "title": "Headline",
            "type": "string"
          },
          "household_id": {
            "title": "Household Id",
            "type": "string"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/EvidenceValue"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "household_id",
          "finding_id",
          "headline",
          "check",
          "values"
        ],
        "title": "FindingEvidence",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HouseholdList": {
        "additionalProperties": false,
        "properties": {
          "households": {
            "items": {
              "$ref": "#/components/schemas/HouseholdRow"
            },
            "title": "Households",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "households"
        ],
        "title": "HouseholdList",
        "type": "object"
      },
...
      "IngestResult": {
        "additionalProperties": false,
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "title": "Errors",
            "type": "array"
          },
          "household_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Household Id"
          },
          "status": {
            "enum": [
              "accepted",
              "needs_review"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "household_id",
          "status",
          "errors"
        ],
        "title": "IngestResult",
        "type": "object"
      },
...
      "MemberCard": {
        "additionalProperties": false,
        "properties": {
          "employer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employer"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "numbers": {
            "items": {
              "$ref": "#/components/schemas/MemberNumber"
            },
            "title": "Numbers",
            "type": "array"
          },
          "person_id": {
            "title": "Person Id",
            "type": "string"
          }
        },
        "required": [
          "person_id",
          "name",
          "employer",
          "numbers"
        ],
        "title": "MemberCard",
        "type": "object"
      },
...
      "Overview": {
        "additionalProperties": false,
        "properties": {
          "changes_since_last_year": {
            "items": {
              "$ref": "#/components/schemas/Change"
            },
            "title": "Changes Since Last Year",
            "type": "array"
          },
          "checklist": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItem"
            },
            "title": "Checklist",
            "type": "array"
          },
          "checklist_label": {
            "const": "Flags for review, not advice",
            "default": "Flags for review, not advice",
            "title": "Checklist Label",
            "type": "string"
          },
          "data_quality": {
            "$ref": "#/components/schemas/DataQuality"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "title": "Errors",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "household_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Household Id"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberCard"
            },
            "title": "Members",
            "type": "array"
          },
          "priority": {
            "anyOf": [
              {
                "enum": [
                  "informational",
                  "low",
                  "medium",
                  "high"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "ruleset_version": {
            "default": "2025.1",
            "title": "Ruleset Version",
            "type": "string"
          },
          "schema_version": {
            "const": "1.1",
            "default": "1.1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "findings",
              "no_findings",
              "needs_review"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Summary"
              },
              {
                "type": "null"
              }
            ]
          },
          "tax_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Year"
          }
        },
        "required": [
          "household_id",
          "tax_year",
          "status",
          "priority",
          "summary",
          "members",
          "checklist",
          "changes_since_last_year",
          "findings",
          "data_quality",
          "errors"
        ],
        "title": "Overview",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Quick-look overview of a household for advisors. Checklist answers are flags for review, not advice. Every displayed number carries a check: verified / mismatch / unconfirmed / conflicted / not_checked.",
    "title": "Household Rapid Analysis",
    "version": "1.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/households": {
      "get": {
        "operationId": "households_list_api_households_get",
        "parameters": [
          {
            "description": "Matches household id or member name",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Matches household id or member name",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "informational",
                    "low",
                    "medium",
                    "high"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "findings",
                    "no_findings",
                    "needs_review"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "needs_documents",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Needs Documents"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Households List",
        "tags": [
          "households"
        ]
      },
      "post": {
        "operationId": "household_ingest_api_households_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Household Ingest",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/households/{household_id}/documents": {
      "post": {
        "description": "Body: {\"name\": \"...\", \"textract\": <AnalyzeDocument response>} or {\"name\": \"...\", \"text\": \"...\", \"form_type\": \"w2\"}.",
        "operationId": "document_ingest_api_households__household_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "title": "Household Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentIngestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Document Ingest",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/households/{household_id}/findings/{finding_id}/evidence": {
      "get": {
        "operationId": "finding_evidence_api_households__household_id__findings__finding_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "title": "Household Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingEvidence"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finding Evidence",
        "tags": [
          "households"
        ]
      }
    },
    "/api/households/{household_id}/members/{person_id}": {
      "get": {
        "operationId": "household_member_api_households__household_id__members__person_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "title": "Household Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberCard"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Household Member",
        "tags": [
          "households"
        ]
      }
    },
    "/api/households/{household_id}/overview": {
      "get": {
        "operationId": "household_overview_api_households__household_id__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "title": "Household Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Household Overview",
        "tags": [
          "households"
        ]
      }
    },
    "/api/households/{household_id}/refresh": {
      "post": {
        "operationId": "household_refresh_api_households__household_id__refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "title": "Household Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Household Refresh",
        "tags": [
          "households"
        ]
      }
    },
...
  }
}